algo for efficienty doing things on first pass
This algorithm I did initially in On2. Friend explains that brute force
is viable but could be better, and explained a first pass solution.

Trick is to store only the values that you need. Take advantage of the
fact that stock profit can only be derived by selling stock today in the
future. Not vice versa.